refactor: DB에 저장될 때 Piece, Position을 문자열로 압축하도록 변경
